Attempt to read EFI labels
Since Windows does not read EFI labels at all, we do not get convenient
partitions created, which we can open and read. So we have to manually
read the EFI partition, then use the start+len in the logic to read ZFS
labels.

In future, it might be better to write a separate kernel part which
creates virtual partitions for EFI disks, and then we can just use the
partition names logic like upstream.

For now, encode names as "#start#length#name".