Use ThreadId instead of Thread
It turns out the HANDLE returned by PsCreateThread is a unique handle to
the creator, and will not match the pseudo handles returned by
PsGetCurrentThread. So we change to use ThreadId instead, look it up in
the thread_create() call, and use it current_thread.

However, using PsGetCurrentThreadId in current_thread makes it crash in
taskq_member, so for now, it is left as the old value. This needs to be
researched and fixed.

dsl_pool_sync_context() directly uses PsGetCurrentThreadId() so that
pool imports will work.