Change vdev_disk IO to be async
Work around the hassle of the IoCompletionRoutine being called at a
level we can not use mutexes, we now spawn a new thread that calls the
io_intr() call immediately, but which sleeps waiting for the Event to be
triggered by IoCompletionRoutine.

Brings pool import from 2.6s to 1.4s - so it appears to improve things.