Initial UNMOUNT support
Since the Volume name is somehow wrong, we can't just query the Volumes
and get mountpoints, so we have to fake it a bit by opening the device
and asking for its unique_id (we return dataset name).

Although unmount completes successfully, and the next mount claims to be
OK, the first access of new mount will panic.