Make zvol operations use _by_dnode routines
This continues what was started in
0eef1bde31d67091d3deed23fe2394f5a8bf2276 by fully converting zvols
to avoid unnecessary dnode_hold() calls. This saves a small amount
of CPU time and slightly improves latencies of operations on zvols.

Reviewed-by: Brian Behlendorf <behlendorf1@llnl.gov>
Signed-off-by: Richard Yao <richard.yao@prophetstor.com>
Closes #6058

Porting-notes:

Cleaned up previous style, which was based on OSX. Since we can use
the standard uio call, we should take advantage of that.

Additionally deleted all the symlink code, as that is specific to
OSX.

Ported-by: Jorgen Lundman <lundman@lundman.net>