Major Mounting rework
Rework devicesetup to register as filesystem and use mntmgr

Add dataset property driveletter=X with "on" as 'first available'
driveletter and "off" to mean to use dataset mountpoint.

Datasets now have property "driveletter" which can be set for a mount.
If it is unset (importing UNIX pool) the rootds will logically get
"driveletter=on" to mean, any available drive letter.
Otherwise if set to a letter A-Z, it will use said letter.

Userland's zmount.c now converts the ZFS mountpoint strings to
Windows strings "\??\C:\dir". In the case that a dataset should
get first available driveletter, the kernel does no extra work.

If the dataset property "driveletter" is used, the kernel replies to
SUGGESTED_LINKNAME to pick specified driveletter.

If no driveletter is to be used, kernel sets up a "reparse point" and
mounts the dataset on the directory.

"zfs mount" will print the mounted datasets and their mountpoints.

Handle reparsepoint dirs in vnop_lookup

It appears detecting reparsepoints is the job of the filesystem, so
parsing the input path looking up directories as we descend, we
test each path element for ZFS_REPARSEPOINT flag. If detected
STATUS_REPARSE is returned, among other special requirements to
point it to the correct mount.

Userland no longer creates the empty directory to mount on, as current
logic only works on C:. At the moment, a two-directory-lower mount
will most likely fail as it does not create parent directory.

Still missing, remove named mount/reparsepoint.

Allow only on, off and letter in driveletter prop

Removed userland mkdir again as it only creates in C: