Successful mount again
So we are back to where we started, but now using the mountmgr
correctly.
It ignores the request to mount it on F: - but it does get mounted at
least.

There is still a problem with file_name_information(), with the
2-byte included length. We can reply like:

file_name_information: remaining space 252 str.len 2 struct size 8
* file_name_information: partial name of \Pr

Where it should be just "\" and yet, has an extra 2 chars

The debug.h VERIFY correction fixes the non-ascii garbage in debug
output.