Mountmgr userland returning drive letter
Kernel mounts are now working well enough that userland can query the
mountmgr for the driveletter assigned to the volume, which is a step
forward.

Unmount and exports will BSOD - that is to be fixed

then try userland asking for different mounts, like
C:\POOL\dataset. Kernel can already handle specific driveletter if so
wanted.