Encryption Stability and On-Disk Format Fixes
The on-disk format for encrypted datasets protects not only the
encrypted and authenticated blocks themselves, but also the order and
interpretation of these blocks. In order to make this work while
maintaining the ability to do raw sends, the indirect bps maintain a
secure checksum of all the MACs in the block below it along with a few
other fields that determine how the data is interpreted.

Unfortunately, the current on-disk format erroneously includes some
fields which are not portable and thus cannot support raw sends. It is
not possible to easily work around this issue due to a separate and
much smaller bug which causes indirect blocks for encrypted dnodes to
not be compressed, which conflicts with the previous bug. In addition,
the current code generates incompatible on-disk formats on big endian
and little endian systems due to an issue with how block pointers are
authenticated. Finally, raw send streams do not currently include
dn_maxblkid when sending both the metadnode and normal dnodes which
are needed in order to ensure that we are correctly maintaining the
portable objset MAC.

This patch zero's out the offending fields when computing the bp MAC
and ensures that these MACs are always calculated in little endian
order (regardless of the host system's byte order). This patch also
registers an errata for the old on-disk format, which we detect by
adding a "version" field to newly created DSL Crypto Keys. We allow
datasets without a version (version 0) to only be mounted for read so
that they can easily be migrated. We also now include dn_maxblkid in
raw send streams to ensure the MAC can be maintained correctly.

Fixes ZOL # 6845

Signed-off-by: Tom Caputi <tcaputi@datto.com>

Fix for ZOL # 6916

When performing zil_claim() at pool import time, it is important that
encrypted datasets set os_next_write_raw before writing to the
zil_header_t. This prevents the code from attempting to
re-authenticate the objset_phys_t when it writes it out, which is
unnecessary because the zil_header_t is not protected by either objset
MAC and impossible since the keys aren't loaded yet. Unfortunately,
one of the code paths did not set this flag, which causes failed
ASSERTs during 'zpool import -F'. This patch corrects this issue.

Signed-off-by: Tom Caputi <tcaputi@datto.com>

Now with data in the test pool

Correct readonly during mount

Raw sends must be able to decrease nlevels

Currently, when a raw zfs send file includes a DRR_OBJECT record
that would decrease the number of levels of an existing object,
the object is reallocated with dmu_object_reclaim() which
creates the new dnode using the old object's nlevels. For non-raw
sends this doesn't really matter, but raw sends require that
nlevels on the receive side match that of the send side so that
the checksum-of-MAC tree can be properly maintained. This patch
corrects the issue by freeing the object completely before
allocating it again in this case.

This patch also corrects several issues with dnode_hold_impl()
and related functions that prevented dnodes (particularly
multi-slot dnodes) from being reallocated properly due to
the fact that existing dnodes were not being fully cleaned up
when they were freed.

Signed-off-by: Tom Caputi <tcaputi@datto.com>
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>

Change os->os_next_write_raw to work per txg

Currently, os_next_write_raw is a single boolean used for determining
whether or not the next call to dmu_objset_sync() should write out the
objset_phys_t as a raw buffer. Since the boolean is not associated
with a txg, the work simply happens during the next txg, which is not
necessarily the correct one. In the current implementation this issue
was misdiagnosed, resulting in a small hack in dmu_objset_sync() which
seemed to resolve the problem.

This patch changes os_next_write_raw to be an array of booleans, one
for each txg in TXG_OFF and removes the hack.

Signed-off-by: Tom Caputi <tcaputi@datto.com>