Handle named mounts
Now the kernel expects all mountpoint strings to be of the form
"\??\C:\POOL\DATASET" for a named mount, or use
"\??\F:" to specify a drive letter directly, or use
"\??\?:" to ask for any available drive letter.

Userland logic will handle the conversion of ZFS (Unix) path to
the expected Windows path, in zmount.c, to come.

Also missing is removing reparse-point on unmount

Also missing is unlinking drive-letter on named mount, currently
it will get both.