Add dataset property driveletter=X
Datasets now have property "driveletter" which can be set for a mount.
If it is unset (importing UNIX pool) the rootds will logically get
"driveltter=?:" to mean, any available drive letter.
Otherwise if set to a letter A-Z, it will use said letter.

If driveletter exists but not set, it will default to Unix mountpoint,
where it looks up the parent's driveletter, then appends dataset path.

Currently only sees immediate parent's drive letter, this will be fixed.

Currently kernel does not remove drive letters, so it unfortunately
finds driveletters where it should not. To be fixed.