Handle reparsepoint dirs in vnop_lookup
It appears detecting reparsepoints is the job of the filesystem, so
parsing the input path looking up directories as we descend, we
test each path element for ZFS_REPARSEPOINT flag. If detected
STATUS_REPARSE is returned, among other special requirements to
point it to the correct mount.

We let userland create the empty directory to mount on, and we
delete it immediately prior to setting the reparse point. This allows
all userland logic to remain as upstream.