Kernel errno and userland did not match
The errno values between userland and kernel were not always agreeing,
which meant ENOTSUP would not be recognised and cause abort().

We could look at including the userland errno.h in the kernel, but it is
unlikely that is allowed.

Related to issue #36 where setting compression=lz4 should have returned:

cannot set property for 'tank': pool and or dataset must be upgraded to set this property or value

instead of calling abort().

However, why the pool thinks it is missing features is the real problem.