Add basic memory pressure events
We relied on that "free mem" was previously 0, thereby making SPL
enter low-memory mode and constantly deflate ARC.

Now we pretend there is plenty memory available, but listen for
low-memory notification from the kernel, and deflate if that is
received.

On my 2G RAM VM the ARC increases to about 220MB, before deflating to
~170MB and growing again, staying between those two limits. The
computed ARC target is about 800M, so it suggests it is doing something
right.