Complete bsd_timeout() bsd_untimeout()
Change the illumos type "int32_t kmem_reaping;" to be a new
bsd_timeout_wrapper structure, wich still starts with same "int32_t".
This allows the helper functions to pass it around like "void *" still.

Under the hood, we have to create a new thread, which is enter
sleep for the desired duration, then call the original function.

Initially tried to use KeSetTimer() with DPC callback, but DPCs are
rather useless, and you can not call any blocking functions (like that
of taskq_dispatch() that kmem_update() tries to call.

Unfortunately, KeCancelTimer() will not signal any WaitForSingleObject()
waiting on it, so those threads would never quit. So instead of calling
KeCancelTimer() we reset the timer back to +1 relative time, so it
triggers immediately.

Luckily, there are only 3 consumers of bsd_timeout() and they could be
re-written to fit better with Windows should someone be eager.

For now, it is pleasing that API conformance could be reached with
bsd_timeout() and bsd_untimeout(), with just the small change of the
type used for variable.