Deletion corrections
Cleanup the file deletion code with regards to handling fileobject. We
make sure not to clear it in vnode before calling zfs_vnop_reclaim, so
zfs itself can handle clearing fileobject->FsContext.

Deleting directories appears to have not worked, in that it will
be removed in Explorer, but if you push F5 it comes back. Turns out
we need to return ENOTEMPTY quite early, in SetDisposition call to
delete directory.