Delete files on CLOSE
Tag files as deleted in CLEANUP and handle the actual deletion in
CLOSE, this lets zp/vp to live longer as required by windows.

We should not list deleted entries in readdir, to be fixed.

We can still die in FltStream, this appears to be because we
release directory memory - if we never release directory vp, we do
not appear to crash - possibly there is an additional conditional
for directories we need to test for.