Milestone: Implement ZVOL support
Attempt to initialise miniport storport driver to handle the virtual
disk creation for ZVOL support. As you can not have your own
MajorFunction callbacks as well as use miniport, we do a hacky copy of
the MajorFunction callback that miniport registers. Then call miniport
from our dispatcher if we do not recognise the IRP.

currently call zvol_open() from create_minor and it remains open the
entire time. There appears to be no open/close semantics when dealing
with disks, we just get read/write requests directly. So we will consider it
"open" as long as Storport has it.

As number of Targets is currently static, we keep an array of 'zv' ptrs
to find the TargetId, which beats searching the list of zvs to find the
correct one.

Currently we do not use LUNs.

Implement multiple ZVOLs and fetch correct zv for IO.

Handle ZVOL volblocksize as preferred. However, I have been unable to
complete "formatting" to NTFS with volblocksize >= 16384. Unsure why, so
changed the ZVOL default volblocksize to 4096 - so simple creates work.

handle zfs destroy zvol

handle zfs export with zvols.