Handle bad paths better
Using winrar to unzip files to ZFS highlighted some issues.

Creating "/dir/file" when "/dir" did not exist (finalpartmaynotexist=1)
did not detect the difference between "file" not existing, and "/dir"
not existing - so it would create a file called "/dir". Now
finalpartmaynotexist also checks that it is the finalcomponent we are
looking at. Returns ESRCH.

Appears to be some code optimisation in effect, it tries to open
"/dir/file", and if it gets error STATUS_OBJECT_NAME_NOT_FOUND then the
file is not there (but it is implied that "/dir" exists.)
If you get STATUS_OBJECT_PATH_NOT_FOUND then it is "/dir" that is
missing. Return the appropriate error.