Advance ZVOL code
As number of Targets is currently static, we keep an array of 'zv' ptrs
to find the TargetId, which beats searching the list of zvs to find the
correct one.

Currently we do not use LUNs.

Implement multiple ZVOLs and fetch correct zv for IO.

Handle ZVOL volblocksize as preferred. However, I have been unable to
complete "formatting" to NTFS with volblocksize >= 16384. Unsure why, so
changed the ZVOL default volblocksize to 4096 - so simple creates work.

handle zfs destroy zvol

handle zfs export with zvols.

this concludes bare-minimal-support for ZVOLs.