Correct spl-condvar clearing of broadcast.
At the moment, cv_destroy() will wait for waiters == 0 before
continuing, however, any waiters we may have sitting in
KeWaitForMultipleObjects() will call KeClearEvent() if they were the
lastwaiter. We have a race then which, cv_destroy() will return and free
the condvar, and lastwaiter calls KeClearEvent on freed memory.

We now check for firstwaiter instead, and clear the event on a known to
be allocated condvar.