Return name error instead of path
When Save Target As in used in Edge browser, it will fail to save to
"\\file" due to us returning STATUS_OBJECT_PATH_NOT_FOUND, ie, that
"\\" not found. We will instead return NAME_NOT_FOUND indicating that
the file is not present.

We probably need to tell the difference between the two, but for now
NAME_NOT_FOUND is less disruptive, since presumably "\\" is always
there.