kmem_alloc to handle size == 0
Upstream handles allocations of size==0, and if supplied KM_SLEEP, we
can not simply return NULL.

This is modelled after Linux's kalloc which returns a "(void *)16"
pointer on size == 0, so that it is a valid address (not NULL) and any
read/write access to the ptr will page-fault. Which is correct, as they
asked for no-memory.

kmem_free will handle special pointer, as well as, NULL in the free
call.

These allocs come from recent crypto, and zfs recv, for which there are
PRs to correct. This commit makes us behave like upstream while waiting
for additional fixes.