Too many ioctl wrappers
On OsX we need to wrap the call to ioctl() with internal versions, like
zioctl() to pass along the real error code. But Windows is able to
define ioctl() itself which handles it, so zioctl() is one wrapper too
many, and we lose the errorcode.

This meant calls like loading crypto key would not return the error
signalling incorrect passphrase used.