SPL: Internal hostid
As not all Registry information is available on boot, let's make
hostid be entirely internal to ZFS. On SPL startup, it will query
the ZFSin Registry for "hostid" and if not set, assign random value.

Bonus side effect to this, is that users can control the hostid
setting desired with RegEdit or similar tools. Stored under
\HLM\SYSTEM\ControlSet001\Services\ZFSin
  hostid  REG_DWORD 0x12345678

We should probably add all kstat tunables to the registry as a way
to set values for boot. (kstat -w for dynamic changes will remain)