Replace fast_mutex with SynchronizationEvent
As pointed out in the ntdev forum, fast_mutex can not be used the way we
were. Here is a simple SynchronizationEvent replacement.

It uses the idea of a mutex being available-to-lock when it is Signaled,
and calls KeWaitForSingleObject() to 'enter/lock' the mutex. If mutex is
'Signaled' it will return immediately, and the mutex is held.
If mutex is already held elsewhere, KeWaitForSingleObject() will block.

mutex_exit() will release the mutex, and set the SynchronizationEvent to
'Signaled' again, which releases one of the waiters, and Clears the
event again.

Speed up mutex by using CAS

Attempt to avoid calling KeWaitForSingleObject() by using atomic
CompareAndSwap to replace the value of m_owner (compared with NULL) and
the value of current_thread(). Only if this fails, do we call the
heavier KeWaitForSingleObject() to block.

Strip OsX mutex leak code

As it had not been ported, and should we want it implemented again
it should be done using getpcstack() style rather than using
__FILE__ style calling convention.