Calling direct IO instead of async with threads
Instead of using a taskq to wait up on IO completion, call teh
ZwWriteFile/ZwReadFile methods directly and block until IO complete.

Quick performance testing showed that this did slightly worse as
not as many IOs are setup to run in parallel. Keeping around
in case we want to test it further.