Mount race and incorrect unmount
It was possible to get a deadlock during mount race, for example when
importing "tank" with dataset "tank/lower", as "tank" was not yet
considered mounted, the mount for "tank/lower" would wait forever. This
was corrected with VPB->Flags |= MOUNTED;

Secondly, if MountMgr is asked to DELETE_POINTS on something that is not
the symlink, it will ignore any future ARRIVAL notifications. So once
something had been mounted, it could not be mounted again. Now we ensure
to only DELETE the driveletter.