Re-port userland (#174)
Instead of littering source file with #ifdef _WIN32, we create
a wrapper for the POSIX calls that uses "int fd;". As we need to
use HANDLEs to open devices, we implement alternate versions of the
POSIX functions that take HANDLEs instead.

Currently the implementation uses simple cast between HANDLE and int,
should we find that the missing bits (64bit -> 32bit) to be a problem
the shim layer could further be enhanced to handle this case.

The functions that have a whole WIN32 replacement are left to use
WIN32 specific code, but as much as possible remains unchanged from
upstream.