The ZFS driver requests memory by NonPagePool, but according MSDN description : https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/wdm/ne-wdm-_pool_type See constants paragraph: NonPagedPool The Microsoft recommends to use NonPagedPoolNx after Windows 8. Some of the places in driver were changed to request from the NonPagedPoolNx All others use NonPagedPool For more information about NonPagePoolNx see: https://docs.microsoft.com/en-us/windows-hardware/drivers/kernel/no-execute-nonpaged-pool
All NonPagePool were changed with NonPagePoolNx.

The function CreateReparsePoint was changed. The type of a parameters SubstituteName and PrintName were changed from LPCWSTR to UNICODE_STRING. All internal calculations was replaced with structure field Length and use buffer directly.

Check the IRQL level and do not call FsRtlEnterFileSystem() if it is greater than APC_LEVEL.
Do not use the Irp pointer if STATUS_PENDING is returned by the dispatcher’s subroutine.