Fix stat_lx mode, change umode default, enable ECP
The #ifdef for ECP inclusions were not working.
file_stat_lx_information uses the mode field to compute permissions, so
translating between ZFS/POSIX mode flags to LX was required.

Change the defuault permission of mkdir to 0777, which is more useful in
Windows. This allows non-root users in lx/ubuntu to use ZFS immediately.