Reimplement CcMgr framework
Prior code would initialise CcMgr once per vnode, but it needs to be
done for each new fileobject, and there are often many fileobjects
pointing to one vnode. Likewise, CcUninitizeCacheMap() needs to be
called for each CLOSE.

Pass in the zp->z_size when opening a new files, calling
vnode_couplefileobject(), so that the correct size can be supplied to
CcInitializeCacheMap().

Implement vflush, to iterate all known vnodes, and each fileobject in
the AVL tree, to forcefully call CcUninitialiseCacheMap() and wait for
all CLOSE to come in, and vnodes drain out.