Allocate CCB with all FileObjects
Prior ito this we would only allocate zfs_dirlist_t for directories. However
as many fileobjects can point to the same vnode (think: many names with
hardlinks) and DeleteOnClose is set, we set the vnode to delete, which
would simulate all-hard-links as deleted.

Turns out the reason other file systems (fastfat) use
fileobject->FsContext2 is for this reason, each FileObject allocates a
CCB unique to that fileobject, and they store information related to just
the fileobject in there (and not in vnode / FsContext as it would affect
all fileobjects).

We now store DeleteOnClose in the CCB/zfs_dirlist_t struct, so that
deleting a single hard-link will only delete it, and not any siblings.

This means the allocation of ccb can be moved into
zfs_couplefileobject() and zfs_decouplefileobject() which makes it even
cleaner.

One difference to fastfat is that we release CCB sooner, in
IRP_MJ_CLEANUP, when they release it in IRP_MJ_CLOSE. It is unclear if
that matters.

The call zfs_setunlink() is then changed to take fileobject instead of
vnode, so we can test for DeleteOnClose.