Port zfs diff
zfs uses pipe() (socketpair under windows) and a few careful
considerations with that. We need to solve the fdopen() problem here,
since zfs send needs it to work one way, but here it actually is using
the FILE* with fprintf. Change libzfs_sendrecv.c to use #if WIN32
instead, and this has a real port of fdopen().

winsock2 will at times use FILE_FLAG_OVERLAPPED, so we must supply an
overlapped structure when issuing IO.

Winsock needs to be initialised, and will automatically do so in
socketpair. We may want to revisit this in future and handle it
differently.