Do not extend file prematurely
The calls to zfs_write() used to extend the file before sending data
to CcMgr - which in turn would call zfs_write to eventually write it
out.

However, if the system would crash, we would be left with a partial
file with the end zero-filled, simply resuming the writes would create a
corrupt file.

We now only tell CcMgr to extend the file, but let the file grow
naturally in the zfs_write() calls issued by CcMgr. Partial files left
on disk are now "correct" up until the point of the crash.