SSV-18196 - BSOD when SCSI RESET completes SRBs concurrently (#211)
Problem: SRBs that are under processing in the zvol_read/write routines cause BSOD when an SRB_FUNCTION_RESET_xxx is processed

Cause: A deprecated StorPort routine would blindly complete all SRBs that comply with the reset criteria even though those were being accessed by the zvol logic.

Fix: Link list all SRBs at the time they are queued in the work item queue. All cancellable SRBs are in that queue. When the reset request comes in mark all linked SRBs conforming to the criteria as "cancelled".
The SRB is taken off the list when the work item is finally processed, checked if "cancelled" and if so complete it with SRB_STATUS_BUSY.

Additional fixes:

Miniport service IRPs were not processed correctly: one IRP in pReverseCallIrp could be forever lost and the final IRP completion did not properly initialize the IoStatus.

ref: SSV-18196