zvols don't appear in disk management (#223)
A previous commit to allow for SRB cancellation did not initialize the Cancelled SRB flag.  That caused a random value there to erroneously treat the I/O as cancelled and SCSI BUSY returned.
This led to disk management waiting a long time (or forever) before it could complete its disks view when zvols were presented.