Go beyond 16 zvols presented to disk manager
Opened up the zvol LUN presentation through StorPort to be able to announce up to 32640 zvols.
By default the maximum amount of ZVOL LUNs is set to 400 (from 16 today) and customization through kstat will be added through a future change.
Also by default the distribution to the targets is set to 32 LUNs per target.  This number will limit the affected LUNs when a SCSI target reset is issued.
The SCSI Report LUNs has been fixed in order to return a list of more than 1 LUN.
The MPIO support has been fixed to take into consideration the bCombineVirtDisks parameter and by
default it is (and remains) FALSE.  This means that, no matter the amount of HBA extensions that are called by StorPort through FindAdapter,
only 1 adapter will announce LUNs and service SCSI requests, and only 1 path will show up in MPIO.
When set to TRUE all adapters defined will respond to SCSI Inquiry and Report Luns, and will therefore enable those LUNs for multipathing in MPIO.
At that point the Zvol disk will show as many adapter paths as defined, set by default to failover-only with only 1 path active/optimized and others in Standby.  So when setting this parameter to TRUE
users should first define the "OpenZFS WinZVOL" VendorId/ProductId in the MPIO console so they can be grabbed and bundled by MPIO.
Because ZFSin's StorPort interface is a virtual miniport, MPIO support is somewhat moot so this parameter should remain FALSE.