Experiment with zfs_fd_t inplace of "int fd"
Since Windows HANDLE type is a 64bit "pointer" we need to be able to
have a larger type for the passing of "fd" around.

Kernel passing was changed to use nvlist_uint64, and rely on
cast truncation to ints. Similar to how zv_cookie is done. (although
new lzc is presumably trying to move away from that)

Experiemented with nvlist_byte_array but was not happy with lack of
kernel safety checks of available space. Did I misunderstand the API?

Another option would be to #define the type just like we set
zfs_fd_t and ZFS_FD_UNSET to have a ZFS_FD_NVLIST_SIZE.

Lastly, this set of changes is large and unattractive. It seems
less certain that upstream would be enthusiastic about it. One
could make a wosix style layer that passes "int fd" around and
lookup the real HANDLE from internal array (pretty much what proc_t's
fd array is) - but that does not handle how to pass them to the kernel.