Mark disks readonly/offline when ZFS has them in-use
Which will stop the simple mistake of partitioning/formatting the wrong
disk. Note that users can still override it and corrupt the disk, but
similarly, users can dd to a raw disk in Unix.

We can decide if this is enough protection from the Users, or if we
should revisit this with more protection.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>