issue#269 - ZFS needs to open disks exclusively
- Positioned the vdev device object accessed below the partition manager or a custom protection filter driver.
- Added an optional custom protection filter driver parm in the registry (zfs_vdev_protection_filter) that a user can set.
To be detected in the device stack the filter driver name takes the form "\Driver\partmgr", for example.

When the vdev is opened the vdev device object that will be accessed for all I/O operations is set to be just below one of the 2 protection filters, whichever is discovered first.

(fixes #269 / ref.SSV-18970)