Add userland variant to vdev_file.c for zdb
vdev_file.c is used by kernel and userland, in the latter
everything is opened as a file, including a disk device.

This gets us close to:

zdb -p //?/C:/src -e tank

working (assuming a file pool is in C:/src/)
but currently gets asserted due to unaligned free.