Fix import partition offset logic
The old code would parse out the start offset of the virtual partition,
then call SetFilePointerEx() to this position. The pread() function
would do offset based seeks from CURRENT, ie, the start of the virtual
partiton, plus offset.

When zdb was corrected, pread() was changed to absolute offset, which
broke import. Now we pass the virtual partition offset into label read
so that it can add on the virtual partition offset, as well as the
label offset within it.

This is only required as we use virtual partitions
(#size#offset#/device/path). If we eventually get the code to make
Windows understand EFI partitions, and create PhysicalDrive/Partition
symlinks for each partition, we can undo the virtual partition hack.