zfs-207 - ensure completion routine call during DI failures
problem: It was possible for a DI request to be failed before being queued up for processing in the ZFS driver. when this happened the pool driver would never get its IRP completion routine called.

fix: if the I/O processing is failed in ZFS prior to queueing to the I/O, added the call for the DI callback to run.
Once the I/O is processed by the queueing mechanism it will always call the DI callback.