ssv-19161 - BSOD: ZFSin!uio_create+0x2f invoked at DISPATCH from DcsPool
Problem: a mutex is taken in the uio_xxxx routines and BSOD the system when DcsPool calls in the DI at DISPATCH level.

Reason: A lot of ZFS routines rely on mutexes instead of spinlocks.  It is fine for DcsPool to be calling in at DISPATCH but ZFS must handle it by not using any of those routines.

Fix: Always queue a workitem, do all the work there and always return STATUS_PENDING to the DI client (i.e.DcsPool). That will take care of all cases now.